+++ vera/__init__.py
@@ -1,4 +1,4 @@
 """Vera: a programming language designed for LLMs."""
 
-__version__ = "0.0.186"
+__version__ = "0.0.187"
 version = __version__